Create README.md
I am a proactive, sociable person and I am always motivated to meet goals and learn more. I appreciate being part of work environments based on respect, teamwork with collective motivation and assertive communication.

Thanks to my journey of almost 4 years providing support in the customer experience area of ​​different companies, I am able to easily solve problems and generate a positive experience for users / clients. My personal challenge at every opportunity is to achieve added value to the expected results.

Regarding my studies, it should be noted that I acquired training in agile methodologies, deepening in Scrum and also in project management. In this approach I found the enthusiasm for managing resources, times and possible risks in order to achieve successful results when carrying out projects aimed at all scales and types of organization.

I believe that the challenge of these times is to think about business strategies that allow companies to position themselves with competitive advantages in the market and at the same time contribute in a sustainable way to the construction of a better society.